refactor: extract message history parser from scheduler
